Move DurationTiming to common
