+++ client/Combatant/Tag.ts
@@ -1,8 +1,8 @@
 import * as ko from "knockout";
 
+import { DurationTiming } from "../../common/DurationTiming";
 import { SavedTag } from "../Encounter/SavedEncounter";
 import { Combatant } from "./Combatant";
-import { DurationTiming } from "./DurationTiming";
 
 export const StartOfTurn: DurationTiming = "StartOfTurn";
 export const EndOfTurn: DurationTiming = "EndOfTurn";
